pe qt: adjusted to the new dg structure, and set v1.7.3
--HG--
extra : convert_revision : svn%3Ac306627e-7827-47d3-bdf0-9a457c9553a1/trunk%4024

+++ pe/qt/build.py
@@ -9,6 +9,7 @@
 # The release version is outdated. Use at least r672 on http://svn.pyinstaller.org/trunk
 
 import os
+import os.path as op
 import shutil
 from app import DupeGuru
 
@@ -17,8 +18,10 @@ def print_and_do(cmd):
     os.system(cmd)
 
 # Removing build and dist
-shutil.rmtree('build')
-shutil.rmtree('dist')
+if op.exists('build'):
+    shutil.rmtree('build')
+if op.exists('dist'):
+    shutil.rmtree('dist')
 
 version = DupeGuru.VERSION
 versioncomma = version.replace('.', ', ') + ', 0'
